feat(hive-btle): Add Peripheral data model for sub-tier devices (#433) (#434)
Implements Peripheral data model in hive-btle for soldier-attached devices
like M5Stack Core2, supporting the HIVE hierarchy: Zone → Cell → Node → Peripheral.

## hive-btle Changes

### New Types (sync/crdt.rs)
- `PeripheralType` enum: SoldierSensor, FixedSensor, Relay
- `EventType` enum: Ping, NeedAssist, Emergency, Moving, InPosition, Ack
- `PeripheralEvent` struct with encode/decode (9 bytes)
- `Peripheral` struct with encode/decode (34-43 bytes)
- 12 unit tests for new types

### ESP32 Compatibility Fixes
- `mesh/manager.rs`: AtomicU64 → AtomicUsize for 32-bit platforms
- `platform/esp32.rs`: Remove Bluedroid imports (using NimBLE)

## M5Stack Core2 Demo (Build 13)

### New Features
- Tap cycles through event types with color-coded display
- Event colors: Green=Ping, Cyan=Moving, Blue=InPosition, Yellow=NeedAssist, Red=Emergency
- Battery updates Peripheral health status
- Backwards-compatible wire format (counter first, peripheral after 0xAB marker)

### Build Fixes
- Use default partition table (single-app) for simpler builds
- Exclude ESP32 example from workspace (requires ESP toolchain)

+++ hive-btle/src/mesh/manager.rs
@@ -7,7 +7,7 @@ use alloc::{boxed::Box, collections::BTreeMap, vec::Vec};
 #[cfg(feature = "std")]
 use std::collections::HashMap;
 
-use core::sync::atomic::{AtomicU64, Ordering};
+use core::sync::atomic::{AtomicUsize, Ordering};
 
 #[cfg(feature = "std")]
 use std::sync::RwLock;
@@ -66,7 +66,8 @@ pub struct MeshManager {
     /// Event callbacks
     callbacks: RwLock<Vec<TopologyCallback>>,
     /// Monotonic time in milliseconds (for testing without system time)
-    current_time_ms: AtomicU64,
+    /// Using AtomicUsize for 32-bit platform compatibility (ESP32)
+    current_time_ms: AtomicUsize,
 }
 
 #[cfg(feature = "std")]
@@ -84,7 +85,7 @@ impl MeshManager {
             candidates: RwLock::new(Vec::new()),
             state: RwLock::new(ManagerState::Stopped),
             callbacks: RwLock::new(Vec::new()),
-            current_time_ms: AtomicU64::new(0),
+            current_time_ms: AtomicUsize::new(0),
         }
     }
 
@@ -149,13 +150,16 @@ impl MeshManager {
     }
 
     /// Set the current time (for testing or embedded without RTC)
+    /// Note: Uses usize internally for 32-bit platform compatibility
     pub fn set_time_ms(&self, time_ms: u64) {
-        self.current_time_ms.store(time_ms, Ordering::SeqCst);
+        self.current_time_ms
+            .store(time_ms as usize, Ordering::SeqCst);
     }
 
     /// Get the current time
+    /// Note: Returns u64 but internally stored as usize for 32-bit compatibility
     pub fn time_ms(&self) -> u64 {
-        self.current_time_ms.load(Ordering::SeqCst)
+        self.current_time_ms.load(Ordering::SeqCst) as u64
     }
 
     /// Get a snapshot of the current topology
